requirements.txt
anyio==3.6.2
APScheduler==3.10.1
asn1crypto==1.5.1
asyncio==3.4.3
beautifulsoup4==4.12.2
bit==0.8.0
blinker==1.6.2
bs4==0.0.1
cairocffi==1.6.0
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
coincurve==18.0.0
cssselect2==0.7.0
defusedxml==0.7.1
exceptiongroup==1.1.2
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Mail==0.9.1
flask-optional-routes==1.4.1
futures==3.0.5
goslate==1.5.4
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanize==4.7.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
nostril @ git+https://github.com/casics/nostril.git@fbc0c91249283a9fbc9036206391ce1138826fd3
packaging==23.1
Pillow==10.0.0
plac==1.3.5
pluggy==1.2.0
prisma==0.9.1
pycparser==2.21
pydantic==1.10.11
pytest==7.4.0
python-bitcoinlib==0.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tabulate==0.9.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.3.6
